[mujoco] rename the files for mujoco robot hardware interface
